requirements.txt
Flask==3.0.2
Flask-Caching==2.3.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
celery==5.0.1
redis==5.0.4
python-dotenv==1.0.1
pdfkit==1.0.0